core team reviews governance document
this is kind of arbitrary, but it seems like a better balance rather
than having core team solely focused on RFC review while having
community team focus on content creation, ecosystem growth, *and* such a
cross-cutting concern as governance. arguably this governance model
should be an RFC anyway, but it feels more accessible as a living
document.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>