give brand repo to community team
this repo just needs a home at the moment. it's a bit outdated and could
use some tidying up but it does at least house some official brand
assets, which may be useful for content that we publish.

if/when a visual design/ux/brand/whatever team is formed it can be moved
there instead.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>