rename 'greenpeace' to 'infrastructure'
i renamed the repo manually because there doesn't seem to be a great way
to rename repos with terraform

i think this might actually fail when 'terraform apply' runs because
it's going to try to archive the 'greenpeace' repo which no longer
exists

or, it'll archive the infrastructure repo through some tragic logic to
follow through to redirected/renamed repos. if that happens i can just
unarchive it, at least.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>