harmonize: fix calculation of role order
for some reason i had it in my head that priority is the inverse of
position, but they're actually the same.

also update the discord client to actually update the position field of
the role object to match its position in the list. it's unclear if the
discord API is based on the order of the array or this field, but the
docs mention the field, so let's set it for good measure.

Signed-off-by: Alex Suraci <asuraci@vmware.com>