Remove repos that were moved to the vmware-archive org
Not sure what process in vmware did this, or why it didn't affect all
repos that are currently archived (e.g. baggageclaim is still in the
Concourse org but is archived).

This may have been messing up membership access for new users as well.
One user on Discord, jphan, reported that we kept getting added then
removed from the org. Not sure if this will fix it. One problem at a
time.

Signed-off-by: Taylor Silva <dev@taydev.net>